OXID eShop Community Edition nightly build ver. 4.5.7 rev. 42367

+++ eshop/core/oxmediaurl.php
@@ -19,7 +19,7 @@
  * @package   core
  * @copyright (C) OXID eSales AG 2003-2011
  * @version OXID eShop CE
- * @version   SVN: $Id: oxmediaurl.php 39009 2011-10-04 11:17:15Z vilma $
+ * @version   SVN: $Id: oxmediaurl.php 38536 2011-09-05 09:02:38Z linas.kukulskis $
  */
 
 /**
@@ -90,7 +90,7 @@ class oxMediaUrl extends oxI18n
         if ( $this->oxmediaurls__oxisuploaded->value ) {
             $sUrl = $this->getConfig()->isSsl() ? $this->getConfig()->getSslShopUrl() : $this->getConfig()->getShopUrl();
             $sUrl .= 'out/media/';
-            $sUrl .= basename($this->oxmediaurls__oxurl->value);
+            $sUrl .= $this->oxmediaurls__oxurl->value;
         } else {
             $sUrl = $this->oxmediaurls__oxurl->value;
         }
@@ -108,12 +108,10 @@ class oxMediaUrl extends oxI18n
     public function delete( $sOXID = null )
     {
         $sFilePath = $this->getConfig()->getConfigParam('sShopDir') . "/out/media/" .
-                     basename($this->oxmediaurls__oxurl->value);
+                     $this->oxmediaurls__oxurl->value;
 
-        if ($this->oxmediaurls__oxisuploaded->value) {
-            if (file_exists($sFilePath)) {
-                unlink($sFilePath);
-            }
+        if ($this->oxmediaurls__oxisuploaded->value && file_exists($sFilePath)) {
+            unlink($sFilePath);
         }
 
         return parent::delete( $sOXID );
